Refactor CiT.CLI.Info - break each into their own file
